requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-groq>=0.3.0
langchain-google-genai>=2.0.0
fire>=0.5.0
streamlit>=1.36.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
